+++ .gitignore
@@ -1,0 +1,2 @@
+frontend/node_modules
+backend/node_modules
