+++ backend/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "puppeteer": "^23.10.3"
+  }
+}
